refactors the functions in custom hook
