+++ client/src/components/App/App.js
@@ -1,6 +1,5 @@
-import React, { useState, useEffect } from "react";
+import React from "react";
 import "./App.scss";
-import axios from "axios";
 
 import AllClasses from "../AllClasses/AllClasses";
 import Navbar from "../Navbar/Navbar";
@@ -18,9 +17,7 @@ import Villager from "../Villager/Villager";
 import Up from "../Up/Up";
 import TeaserPage from "../TeaserPage/TeaserPage";
 import GuestProfile from "../GuestProfile/GuestProfile";
-
-// import openSocket from 'socket.io-client';
-// import ChatWindow from '../ChatWindow/ChatWindow';
+import { useAppData } from "../../hooks/useAppData";
 
 const LOGIN = "LOGIN";
 const REGISTER = "REGISTER";
@@ -32,258 +29,19 @@ const EDIT = "EDIT";
 const LOADING = "LOADING";
 const VILLAGER_QUESTS = "VILLAGER_QUESTS";
 const TAKEN = "TAKEN";
-// const CHAT = 'CHAT';
 const ABOUT = "ABOUT";
 const TEASER = "TEASER";
 const GUEST_PROFILE = "GUEST_PROFILE";
 
 export default function App() {
-  const [state, setState] = useState({
-    classesProgressData: [],
-    classesData: [],
-    userData: {},
-    userQuests: [],
-    villagers: [],
-    badges: [],
-    userBadges: [],
-    questsByVillager: [],
-    questsByAdventurer: [],
-    sessions: 0,
-    adventurer: false,
-    username: "",
-    view: TEASER,
-    loggedIn: false,
-    adventurers: [],
-    guestInfo: [],
-    guestBadges: [],
-    // socket: [],
-    // chatMessages: [],
-    // knownUsers: {},
-  });
-
-  function isEmpty(obj) {
-    return !obj || Object.keys(obj).length === 0;
-  }
-
-  //Get Initial state data
-  useEffect(() => {
-    axios
-      .get("/checkSession")
-      .then((response) => {
-        if (!isEmpty(response.data[0])) {
-          setState((prevState) => {
-            return {
-              ...prevState,
-              userData: response.data[0],
-              view: response.data[0]
-                ? response.data[0].adventurer
-                  ? SHOW
-                  : CREATE
-                : LOGIN,
-              sessions: response.data[0].id,
-              adventurer: response.data[0].adventurer,
-              username: response.data[0].first_name,
-              loggedIn: true,
-            };
-          });
-        }
-      })
-      .catch((error) => console.log(error));
-  }, []);
-
-  useEffect(() => {
-    if (state.loggedIn) {
-      const getUserBadges = axios.get(`/users/${state.userData.id}/badges`);
-      const getUserQuest = axios.get("/quests");
-      const getClasses = axios.get("/classes");
-      const getClassesProgress = axios.get("/userClassProgress");
-      const getVillagers = axios.get("/villagers");
-      const getBadges = axios.get("/badges");
-      const getQuestsByVillager = axios.get(
-        `/users/${state.userData.id}/quests`
-      );
-      const getQuestsByAdventurer = axios.get(
-        `/users/adventurer/${state.userData.id}/quests`
-      );
-      const getAdventurers = axios.get(`/adventurers`);
-      Promise.all([
-        getUserBadges,
-        getUserQuest,
-        getClasses,
-        getClassesProgress,
-        getVillagers,
-        getBadges,
-        getQuestsByVillager,
-        getQuestsByAdventurer,
-        getAdventurers,
-      ])
-        .then(
-          ([
-            { data: userBadges },
-            { data: userQuests },
-            { data: classesData },
-            { data: classesProgressData },
-            { data: villagers },
-            { data: badges },
-            { data: questsByVillager },
-            { data: questsByAdventurer },
-            { data: adventurers },
-          ]) => {
-            setState({
-              ...state,
-              userBadges,
-              userQuests,
-              classesData,
-              classesProgressData,
-              villagers,
-              badges,
-              questsByVillager,
-              questsByAdventurer,
-              adventurers,
-            });
-          }
-        )
-        .catch((err) => console.log(err));
-    } else {
-      const userQuest = await axios.get("/quests").data;
-      const villagers = await axios.get("/villagers").data;
-      const classesData = await axios.get("/classes").data;
-      setState((prevState) => {
-        return {
-          ...prevState,
-          userQuests,
-          villagers,
-          classesData,
-        };
-     });
-    }
-  }, [state.loggedIn]);
-
-  //Socket.io
-  // const addNewMessage = function (msgObj) {
-  //   setState((prevState) => {
-  //     return {
-  //       ...prevState,
-  //       chatMessages: [
-  //         ...prevState.chatMessages,
-  //         {
-  //           message: msgObj.msg,
-  //           userData: prevState.knownUsers[msgObj.userId],
-  //         },
-  //       ],
-  //     };
-  //   });
-  // };
-
-  // const newUserCheck = function (msgObj) {
-  //   if (state.knownUsers[msgObj.userId]) {
-  //     addNewMessage(msgObj);
-  //   } else {
-  //     axios.get(`/users/${msgObj.userId}`).then((results) => {
-  //       setState((prevState) => {
-  //         return {
-  //           ...prevState,
-  //           knownUsers: {
-  //             ...prevState.knownUsers,
-  //             [msgObj.userId]: results.data[0],
-  //           },
-  //         };
-  //       });
-  //       addNewMessage(msgObj);
-  //     });
-  //   }
-  // };
-
-  // const openNewSocket = () => {
-  //   let socket = openSocket('localhost:8081');
-
-  //   socket.on('connect', function () {
-  //     socket.on('chat message', (msgObj) => {
-  //       newUserCheck(msgObj);
-  //     });
-  //     setState((prevState) => {
-  //       return {
-  //         ...prevState,
-  //         socket: socket,
-  //       };
-  //     });
-  //   });
-  // };
-
-  const changeView = (viewType) => {
-    setState({
-      ...state,
-      view: LOADING,
-    });
-    setTimeout(() => {
-      setState({
-        ...state,
-        view: viewType,
-      });
-    }, 500);
-  };
-
-  const handleLogin = (edit) => {
-    return axios
-      .get("/checkSession")
-      .then((response) => {
-        setState((prevState) => {
-          return {
-            ...prevState,
-            userData: response.data[0],
-            sessions: response.data[0].id,
-            adventurer: response.data[0].adventurer,
-            username: response.data[0].first_name,
-            loggedIn: true,
-            view: edit ? PROFILE : response.data[0].adventurer ? SHOW : CREATE,
-          };
-        });
-      })
-      .catch((error) => console.log(error));
-  };
-
-  const handleLogout = () => {
-    return axios
-      .post("/logout")
-      .then(() => {
-        setState((prev) => ({
-          ...prev,
-          classesProgressData: [],
-          classesData: [],
-          userData: {},
-          userQuests: [],
-          villagers: [],
-          badges: [],
-          sessions: 0,
-          adventurer: false,
-          username: "",
-          loggedIn: false,
-          view: LOGIN,
-          adventurers: [],
-          guestInfo: [],
-          guestBadges: [],
-        }));
-      })
-      .catch((error) => console.log(error));
-  };
-
-  const getGuestProfile = function (id) {
-    const getUserInfo = axios.get(`/users/${id}`);
-    const getUserBadges = axios.get(`/users/${id}/badges`);
-
-    Promise.all([getUserInfo, getUserBadges])
-      .then(([{ data: guestInfo }, { data: guestBadges }]) =>
-        setState((prevState) => {
-          return {
-            ...prevState,
-            guestInfo,
-            guestBadges,
-            view: GUEST_PROFILE,
-          };
-        })
-      )
-      .catch((e) => console.log(e));
-  };
+  const {
+    state,
+    getGuestProfile,
+    handleLogout,
+    handleLogin,
+    changeView,
+    setState,
+  } = useAppData();
 
   return (
     <div className="App">
@@ -301,8 +59,6 @@ export default function App() {
           onVillagerQuests={() => changeView(VILLAGER_QUESTS)}
           onTaken={() => changeView(TAKEN)}
           onAbout={() => changeView(ABOUT)}
-
-          // onChat={() => changeView(CHAT)}
         />
       ) : (
         <Navbar
@@ -341,9 +97,6 @@ export default function App() {
             state={state}
             setState={setState}
             onGuestProfile={getGuestProfile}
-            // newUserCheck={newUserCheck}
-            // openNewSocket={openNewSocket}
-            // addNewMessage={addNewMessage}
           />
         )}
         {state.view === PROFILE && (
@@ -356,14 +109,6 @@ export default function App() {
             onProfile={() => changeView(PROFILE)}
           />
         )}
-        {/* {state.view === CHAT && (
-          <ChatWindow
-            socket={state.socket}
-            openNewSocket={openNewSocket}
-            messages={state.chatMessages}
-            loggedInUser={state.userData}
-          />
-        )} */}
         {state.view === VILLAGER_QUESTS && (
           <VillagerQuestList state={state} setState={setState} />
         )}
